pyproject.toml
[tool.poetry]
name = "curs-bot"
version = "0.1.0"
description = ""
authors = ["nuris <ntasbolatov03@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.6.0"
pydantic-settings = "^2.2.1"
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
psycopg = "^3.1.19"
pymorphy2-dicts-ru = "^2.4.417127.4579844"
setuptools = "^70.0.0"
nltk = "^3.8.1"
spacy = "^3.7.5"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
